+++ src/bfabric_scripts/bfabric_upload_submitter_executable.py
@@ -39,7 +39,7 @@ from pathlib import Path
 
 from bfabric import Bfabric
 from bfabric.cli_formatting import setup_script_logging
-from bfabric_scripts.cli.external_job.upload_submitter_executable import upload_submitter_executable_impl
+from bfabric_scripts.cli.external_job.upload_submitter_executable import upload_submitter_executable
 
 
 def main() -> None:
@@ -57,7 +57,7 @@ def main() -> None:
     parser.add_argument("--name", type=str, help="Name of the submitter", required=False)
     parser.add_argument("--description", type=str, help="Description about the submitter", required=False)
     options = parser.parse_args()
-    upload_submitter_executable_impl(client=client, **vars(options))
+    upload_submitter_executable(client=client, **vars(options))
 
 
 if __name__ == "__main__":